+++ sash-libs/sash-err-stack/sash-err-stack.sh
@@ -10,50 +10,36 @@
 # S.A.S.H. is the main way to add things to your ~/.bashrc and still
 # maintain structure.
 
+_sash_initial_err_mode=""
+
+if [[ "$-" =~ "e" ]]; then
+  _sash_initial_err_mode="0"
+else
+  _sash_initial_err_mode="1"
+fi
+
+set +e
+
+declare -f _sash_safe_add_to_trap >/dev/null
+__sash_trap_intermediate_check="$?"
+
+if [[ "$__sash_trap_intermediate_check" == "1" ]]; then
+  if [[ "x$SASH_TRAP_DIR" == "x" ]]; then
+    source "$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )/../sash-trap/sash-trap.sh"
+  else
+    source "$SASH_TRAP_DIR/sash-trap.sh"
+  fi
+fi
+
+unset __sash_trap_intermediate_check
+
+
 declare -f __sash_pop_err_mode_stack >/dev/null
 __sash_intermediate_check="$?"
 
 if [[ "$__sash_intermediate_check" == "1" ]]; then
 
-# _sash_get_trapped_text(signal: String)
-#
-# Modifies Variables: None
-#
-# parses trap output to get you the command for a signal.
-_sash_get_trapped_text() {
-  local signal="$1"
-  echo "$(trap -p "$signal" | sed "s/trap -- '//g; s/' $signal$//g; s/\\\''//g")"
-}
-
-# _sash_safe_add_to_trap(signal: String, command: String)
-#
-# Modifies Variables: None
-#
-# safely adds a command to a trap.
-_sash_safe_add_to_trap() {
-  local command_to_add="$1"
-  local signal="$2"
-
-  if [[ "x$(trap -p "$signal")" == "x" ]]; then
-    trap "$command_to_add" "$signal"
-  else
-    local trapped_text="$(_sash_get_trapped_text "$signal")"
-    echo "$trapped_text" | grep "^.*;" > /dev/null 2>&1
-    if [[ "$?" == "0" ]]; then
-      trap "$trapped_text $command_to_add;" "$signal"
-    else
-      trap "$trapped_text; $command_to_add;" "$signal"
-    fi
-  fi
-}
-
 _sash_global_err_mode_stack=()
-_sash_initial_err_mode=""
-if [[ "$-" =~ "e" ]]; then
-  _initial_err_mode="0"
-else
-  _initial_err_mode="1"
-fi
 
 # __sash_reset_initial_stack()
 #
@@ -79,10 +65,9 @@ __sash_reset_initial_stack() {
 #
 # Pushes onto the error stack.
 __sash_push_err_mode_stack() {
-  local func_name="$1"
-  local err_mode="$2"
-  _sash_global_err_mode_stack=("${_sash_global_err_mode_stack[@]}" "$func_name|$err_mode")
-  if [[ "$err_mode" == "1" ]]; then
+  _sash_global_err_mode_stack=("${_sash_global_err_mode_stack[@]}" "$1|$2")
+
+  if [[ "$2" == "1" ]]; then
     set -e
   else
     set +e
@@ -99,34 +84,39 @@ __sash_push_err_mode_stack() {
 # Should be called when a function exits. Checks to see
 # if the function exiting had an artificial error_state, and unsets it.
 __sash_pop_err_mode_stack() {
-  local func_name="${FUNCNAME[1]}"
-  local new_arr=()
-  local func_iter=
-  for func_iter in "${_sash_global_err_mode_stack[@]}"; do
-    local func_iter_inner_arr=(${func_iter//|/\ })
-    if [[ "${func_iter_inner_arr[0]}" == "$func_name" ]]; then
-      local was_enabled_err="${func_iter_inner_arr[1]}"
-      if [[ "$was_enabled_err" == "1" ]]; then
-        set +e
-      else
-        set -e
-      fi
-    else
-      new_arr=("${new_arr[@]}" "$func_iter")
+  local readonly func_name="${FUNCNAME[1]}"
+
+  local readonly stack_size="${#_sash_global_err_mode_stack[@]}"
+  local readonly stack_le_index=$(( stack_size - 1 ))
+  if [[ "$stack_size" == "0" ]]; then
+    return 0
+  fi
+
+  local readonly stack_last_element="${_sash_global_err_mode_stack[$stack_le_index]}"
+  local readonly stack_split_name=(${stack_last_element//|/\ })
+  if [[ "${stack_split_name[0]}" == "$func_name" ]]; then
+    if [[ "$stack_size" == "1" ]]; then
+      __sash_reset_initial_stack
+      _sash_global_err_mode_stack=()
+      return 0
     fi
-  done
-  _sash_global_err_mode_stack=("${new_arr[@]}")
-  if [[ "${#_sash_global_err_mode_stack[@]}" == "0" ]]; then
-    __sash_reset_initial_stack
+
+    if [[ "${stack_split_name[1]}" == "1" ]]; then
+      set +e
+    else
+      set -e
+    fi
+
+    _sash_global_err_mode_stack=(${_sash_global_err_mode_stack[@]:0:$stack_le_index})
   else
-    local last_elem="${_sash_global_err_mode_stack[-1]}"
-    local inner_split=(${last_elem//|/\ })
-    if [[ "${inner_split[1]}" == "1" ]]; then
+    if [[ "${stack_split_name[1]}" == "1" ]]; then
       set -e
     else
       set +e
     fi
   fi
+
+  return 0
 }
 
 set -o functrace
@@ -139,8 +129,7 @@ _sash_safe_add_to_trap "__sash_pop_err_mode_stack" "RETURN"
 #
 # sets the error mode to be off for the runtime of this function.
 __sash_allow_errors() {
-  local func_name="${FUNCNAME[1]}"
-  __sash_push_err_mode_stack "$func_name" "0"
+  __sash_push_err_mode_stack "${FUNCNAME[1]}" "0"
 }
 
 # __sash_guard_errors()
@@ -150,9 +139,9 @@ __sash_allow_errors() {
 #
 # sets the error mode to be on for the runtime of this function.
 __sash_guard_errors() {
-  local func_name="${FUNCNAME[1]}"
-  __sash_push_err_mode_stack "$func_name" "1"
+  __sash_push_err_mode_stack "${FUNCNAME[1]}" "1"
 }
 fi
 
 unset __sash_intermediate_check
+__sash_reset_initial_stack