+++ sash-libs/sash-err-stack/tests/test-cases/test-can-be-sourced-in-err-mode.sh
@@ -1,0 +1,9 @@
+#!/usr/bin/env bash
+
+set -e
+
+SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+source "$SCRIPT_DIR/../../sash-err-stack.sh"
+
+false
+echo "hey"