+++ sash-libs/sash-err-stack/tests/test-cases/test-err-mode-mixup.sh
@@ -1,0 +1,19 @@
+#!/usr/bin/env bash
+
+SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+source "$SCRIPT_DIR/../../sash-err-stack.sh"
+
+function messWithGlobalErrorMode() {
+  set +e
+  false
+  echo "hey"
+}
+
+function guard() {
+  __sash_guard_errors
+  messWithGlobalErrorMode
+  false
+  echo "hello world"
+}
+
+guard