+++ sash-libs/sash-err-stack/tests/test-cases/test-source-multiple-times.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+source "$SCRIPT_DIR/../../sash-err-stack.sh"
+source "$SCRIPT_DIR/../../sash-err-stack.sh"