+++ sash-libs/sash-trap/tests/test-cases/test-add-multiple-to-empty-trap.sh
@@ -1,0 +1,16 @@
+#!/usr/bin/env bash
+
+SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+source "$SCRIPT_DIR/../../sash-trap.sh"
+
+function onTrap() {
+  echo "onTrap"
+}
+
+( \
+  trap - "SIGALRM" && \
+  _sash_safe_add_to_trap "onTrap" "SIGALRM" && \
+  _sash_safe_add_to_trap "onTrap" "SIGALRM" && \
+  _sash_safe_add_to_trap "onTrap" "SIGALRM" && \
+  kill -s SIGALRM $BASHPID \
+)